--- android/build.gradle
@@ -1,26 +1,0 @@
-apply plugin: 'com.android.library'
-
-android {
-    compileSdkVersion 26
-    buildToolsVersion "26.0.3"
-    defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 26
-        versionCode 1
-        versionName "1.0"
-
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-
-    }
-
-    buildTypes {
-        release {
-
-        }
-    }
-
-}
-
-dependencies {
-    compile 'com.facebook.react:react-native:+'
-}
