--- android/src/main/AndroidManifest.xml
@@ -1,26 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.puti.upgrade">
-
-    <uses-permission android:name="android.permission.INTERNET" />
-    <!--8.0安装需要的权限-->
-    <uses-permission android:name="android.permission.REQUEST_INSTALL_PACKAGES" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
-    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
-
-    <application>
-        <provider
-            android:name=".CustomFileProvider"
-            android:authorities="${applicationId}.updateFileProvider"
-            android:exported="false"
-            android:grantUriPermissions="true">
-            <meta-data
-                android:name="android.support.FILE_PROVIDER_PATHS"
-                android:resource="@xml/update_file_provider" />
-        </provider>
-
-        <service
-            android:name=".DownloadService"
-            android:exported="true" />
-    </application>
-
-</manifest>
