--- android/src/main/java/com/puti/upgrade/AppUtils.java
@@ -1,29 +1,0 @@
-package com.puti.upgrade;
-
-import android.content.Context;
-import android.content.pm.PackageManager;
-
-
-public class AppUtils {
-
-    public static int getVersionCode(Context mContext) {
-        if (mContext != null) {
-            try {
-                return mContext.getPackageManager().getPackageInfo(mContext.getPackageName(), 0).versionCode;
-            } catch (PackageManager.NameNotFoundException ignored) {
-            }
-        }
-        return 0;
-    }
-
-    public static String getVersionName(Context mContext) {
-        if (mContext != null) {
-            try {
-                return mContext.getPackageManager().getPackageInfo(mContext.getPackageName(), 0).versionName;
-            } catch (PackageManager.NameNotFoundException ignored) {
-            }
-        }
-
-        return "";
-    }
-}
