--- android/src/main/java/com/puti/upgrade/CustomFileProvider.java
@@ -1,10 +1,0 @@
-package com.puti.upgrade;
-
-import android.support.v4.content.FileProvider;
-
-/**
- * Created by puti on 2018/4/26.
- */
-
-public class CustomFileProvider extends FileProvider {
-}
