--- android/src/main/java/com/puti/upgrade/HttpUtils.java
@@ -1,56 +1,0 @@
-package com.puti.upgrade;
-
-import android.util.Log;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-
-public class HttpUtils {
-
-
-    public static String get(String urlStr) {
-        HttpURLConnection uRLConnection = null;
-        InputStream is = null;
-        BufferedReader buffer = null;
-        String result = null;
-        try {
-            URL url = new URL(urlStr);
-            uRLConnection = (HttpURLConnection) url.openConnection();
-            uRLConnection.setRequestMethod("GET");
-
-            is = uRLConnection.getInputStream();
-            buffer = new BufferedReader(new InputStreamReader(is));
-            StringBuilder strBuilder = new StringBuilder();
-            String line;
-            while ((line = buffer.readLine()) != null) {
-                strBuilder.append(line);
-            }
-            result = strBuilder.toString();
-        } catch (Exception e) {
-            Log.e(Constants.TAG, "http post error");
-        } finally {
-            if (buffer != null) {
-                try {
-                    buffer.close();
-                } catch (IOException ignored) {
-
-                }
-            }
-            if (is != null) {
-                try {
-                    is.close();
-                } catch (IOException ignored) {
-
-                }
-            }
-            if (uRLConnection != null) {
-                uRLConnection.disconnect();
-            }
-        }
-        return result;
-    }
-}
