--- android/src/main/res/xml/update_file_provider.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<paths>
-    <!--升级-->
-    <external-cache-path
-        name="update_external_cache"
-        path="." />
-
-    <cache-path
-        name="update_cache"
-        path="." />
-</paths>
