+++ ios_upgrade/upgrade.m
@@ -1,7 +1,7 @@
 
 #import "upgrade.h"
 
-#import <UIKit/UIKit.h>
+#import "AppDelegate.h"
 #import <React/RCTBridgeModule.h>
